Relocated App logic
